Split computation and display of changes
